pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "manual-chatbot"
version = "0.1.0"
description = "Smart Chunking Pipeline for Vehicle Service Manual RAG"
requires-python = ">=3.10"
dependencies = [
    "pymupdf>=1.23.0",
    "pyyaml>=6.0",
    "qdrant-client>=1.7.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
]

[project.scripts]
pipeline = "pipeline.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = [
    "unit: Unit tests (no external dependencies)",
    "integration: Integration tests (require external services)",
    "slow: Slow tests",
]

[tool.setuptools.packages.find]
where = ["src"]
